made some updates and fixed some bugs

+++ src/services/familyService.ts
@@ -1,5 +1,5 @@
 import { db, storage } from '@/firebase';
-import { collection, addDoc,getDoc, deleteDoc, doc, updateDoc, getDocs,where,query } from 'firebase/firestore';
+import { collection, addDoc,getDoc, deleteDoc, doc, updateDoc, getDocs,where,query, deleteField } from 'firebase/firestore';
 import { ref, uploadBytes, getDownloadURL, deleteObject } from 'firebase/storage';
 import { FamilyMember } from '@/types/familyTypes';
 
@@ -46,7 +46,7 @@ export const addFamilyMember = async (newMember: FamilyMember, image: File | nul
   }
 
   try {
-    const docRef = await addDoc(collection(db, "familyMembersTest"), memberToAdd);
+    const docRef = await addDoc(collection(db, "familyMembers"), memberToAdd);
     if (newMember.pids && newMember.pids.length) {
       await updateFamilyMember(newMember.pids[0], { pids: [docRef.id] });
     }
@@ -59,7 +59,7 @@ export const addFamilyMember = async (newMember: FamilyMember, image: File | nul
 
 export const getTreeNames = async () => {
   try {
-    const rootMembersRef = collection(db, "familyMembersTest");
+    const rootMembersRef = collection(db, "familyMembers");
     const querySnapshot = await getDocs(rootMembersRef);
     const treeNames = querySnapshot.docs.map(doc => doc.data().treename);
     return [...new Set(treeNames)];
@@ -68,27 +68,79 @@ export const getTreeNames = async () => {
   }
 };
 
-export const deleteTree = async (treeName: string) => {
+export const deleteTree = async (treename: string, confirmation: string) => {
+  if (confirmation !== treename) {
+    throw new Error("Tree name confirmation does not match. Deletion aborted.");
+  }
+
   try {
     const rootMembersRef = collection(db, "rootFamilyMembers");
-    const q = query(rootMembersRef, where("treeName", "==", treeName));
+    const q = query(rootMembersRef, where("treename", "==", treename));
     const querySnapshot = await getDocs(q);
 
     if (!querySnapshot.empty) {
       await deleteDoc(doc(db, "rootFamilyMembers", querySnapshot.docs[0].id));
     }
 
-    const familyMembersRef = collection(db, "familyMembersTest");
+    const familyMembersRef = collection(db, "familyMembers");
     const familyMembersSnapshot = await getDocs(familyMembersRef);
-    familyMembersSnapshot.forEach(async doc => {
-      if (doc.data().treename === treeName) {
+    const deletePromises = familyMembersSnapshot.docs.map(async doc => {
+      if (doc.data().treename === treename) {
         await deleteFamilyMember(doc.id);
       }
     });
+    await Promise.all(deletePromises);
+
+    console.log(`Tree "${treename}" has been successfully deleted.`);
   } catch (error) {
     console.error('Error deleting tree:', error);
+    throw error;
   }
 }
+
+export const updateAllFamilyMembersWithTreeName = async () => {
+  try {
+    const familyMembersRef = collection(db, "familyMembers");
+    const querySnapshot = await getDocs(familyMembersRef);
+
+    const updatePromises = querySnapshot.docs.map(async (doc) => {
+      const memberRef = doc.ref;
+      await updateDoc(memberRef, {
+        treename: "Default Tree"
+      });
+    });
+
+    await Promise.all(updatePromises);
+
+    console.log("All family members updated with treename: Default Tree");
+  } catch (error) {
+    console.error("Error updating family members with treename:", error);
+  }
+};
+
+
+export const deleteEmptyIdField = async () => {
+ try {
+    const familyMembersRef = collection(db, "familyMembers");
+    const querySnapshot = await getDocs(familyMembersRef);
+
+    const updatePromises = querySnapshot.docs.map(async (document) => {
+      const docData = document.data();
+      if ('id' in docData && docData.id === null) {
+        const docRef = doc(db, "familyMembers", document.id);
+        await updateDoc(docRef, {
+          id: deleteField()
+        });
+      }
+    });
+
+    await Promise.all(updatePromises);
+
+    console.log("Empty 'id' fields have been deleted from all documents");
+  } catch (error) {
+    console.error("Error deleting empty 'id' fields:", error);
+  }
+};
 
 export const downLoadFamilyMembersToTxt = async () => {
   try {
@@ -102,15 +154,15 @@ export const downLoadFamilyMembersToTxt = async () => {
     name1
     ...
     */
-    const familyMembersRef = collection(db, "familyMembersTest");
+    const familyMembersRef = collection(db, "familyMembers");
     const familyMembersSnapshot = await getDocs(familyMembersRef);
     const familyMembers = familyMembersSnapshot.docs.map(doc => doc.data() as FamilyMember);
 
     const treeNames = [...new Set(familyMembers.map(member => member.treename))];
-    const txtData = treeNames.map(treeName => {
-      const members = familyMembers.filter(member => member.treename === treeName);
+    const txtData = treeNames.map(treename => {
+      const members = familyMembers.filter(member => member.treename === treename);
       const memberNames = members.map(member => member.name);
-      return [treeName, ...memberNames];
+      return [treename, ...memberNames];
     });
 
     const txt = txtData.map(data => data.join("\n")).join("\n\n");
@@ -129,7 +181,7 @@ export const downLoadFamilyMembersToTxt = async () => {
 
 export const updateFamilyMember = async (memberId: string, updates: Partial<FamilyMember>) => {
    try {
-    const memberRef = doc(db, "familyMembersTest", memberId);
+    const memberRef = doc(db, "familyMembers", memberId);
     const memberSnapshot = await getDoc(memberRef);
 
     const currentData = memberSnapshot.data();
@@ -146,7 +198,7 @@ export const updateFamilyMember = async (memberId: string, updates: Partial<Fami
 
 export const setFamilyMember = async (memberId: string, updates: Partial<FamilyMember>) => {
   try {
-    const memberRef = doc(db, "familyMembersTest", memberId);
+    const memberRef = doc(db, "familyMembers", memberId);
     await updateDoc(memberRef, updates);
   } catch (error) {
     console.error("Error setting document: ", error);
@@ -156,8 +208,8 @@ export const setFamilyMember = async (memberId: string, updates: Partial<FamilyM
 
 export const deleteFamilyMember = async (memberId: string) => {
   try {
-    const memberRef = doc(db, "familyMembersTest", memberId);
-    const memberSnapshot = await getDocs(collection(db, "familyMembersTest"));
+    const memberRef = doc(db, "familyMembers", memberId);
+    const memberSnapshot = await getDocs(collection(db, "familyMembers"));
     const memberToDelete = memberSnapshot.docs.find(doc => doc.id === memberId)?.data() as FamilyMember;
 
     if (memberToDelete.img) {
@@ -180,25 +232,25 @@ export const deleteFamilyMember = async (memberId: string) => {
 
 export const clearNoExistantData = async () => {
   try {
-    const querySnapshot = await getDocs(collection(db, "familyMembersTest"));
+    const querySnapshot = await getDocs(collection(db, "familyMembers"));
     querySnapshot.forEach(async (doc) => {
       const member = doc.data() as FamilyMember;
       if (member.fid?.length) {
-        const fatherSnapshot = await getDocs(collection(db, "familyMembersTest"));
+        const fatherSnapshot = await getDocs(collection(db, "familyMembers"));
         if (!fatherSnapshot.docs.find(father => father.id === member.fid[0])) {
           await setFamilyMember(doc.id, { fid: [] });
         }
       }
       if (member.mid?.length) {
-        const q = query(collection(db, "familyMembersTest"), where("capital", "==", true));
-        const motherSnapshot = await getDocs(collection(db, "familyMembersTest"));
+        const q = query(collection(db, "familyMembers"), where("capital", "==", true));
+        const motherSnapshot = await getDocs(collection(db, "familyMembers"));
 
         if (!motherSnapshot.docs.find(mother => mother.id === member.mid[0])) {
           await setFamilyMember(doc.id, { mid: [] });
         }
       }
       if (member.pids?.length) {
-        const partnerSnapshot = await getDocs(collection(db, "familyMembersTest"));
+        const partnerSnapshot = await getDocs(collection(db, "familyMembers"));
         const updatedPartners = member.pids.filter(partner => partnerSnapshot.docs.find(p => p.id === partner));
         const uniquePartners = [...new Set(updatedPartners)];
         if (updatedPartners.length === uniquePartners.length) {
@@ -214,14 +266,14 @@ export const clearNoExistantData = async () => {
   }
 };
 
-export const setRootFamilyMember = async (memberId: string, treeName: string) => {
+export const setRootFamilyMember = async (memberId: string, treename: string) => {
   try {
     const rootMembersRef = collection(db, "rootFamilyMembers");
-    const q = query(rootMembersRef, where("treeName", "==", treeName));
+    const q = query(rootMembersRef, where("treename", "==", treename));
     const querySnapshot = await getDocs(q);
 
     if (querySnapshot.empty) {
-      await addDoc(rootMembersRef, { treeName, rootMemberId: memberId });
+      await addDoc(rootMembersRef, { treename, rootMemberId: memberId });
     } else {
       const docRef = doc(db, "rootFamilyMembers", querySnapshot.docs[0].id);
       await updateDoc(docRef, { rootMemberId: memberId });
@@ -232,10 +284,10 @@ export const setRootFamilyMember = async (memberId: string, treeName: string) =>
   }
 }
 
-export const getRootFamilyMember = async (treeName: string) => {
+export const getRootFamilyMember = async (treename: string) => {
   try {
     const rootMembersRef = collection(db, "rootFamilyMembers");
-    const q = query(rootMembersRef, where("treeName", "==", treeName));
+    const q = query(rootMembersRef, where("treename", "==", treename));
     const querySnapshot = await getDocs(q);
 
     if (!querySnapshot.empty) {
